2.ai-library/pyproject.toml
[project]
name = "knowledge-library"
version = "0.1.0"
description = "Personal Knowledge Library System - The Reliable Librarian"
requires-python = ">=3.11"
dependencies = [
    "claude-code-sdk>=0.0.23",
    "anyio>=4.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.7.0",
    "httpx>=0.26.0",
    # Phase D: REST API
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: requires external services (e.g., Qdrant, API keys)",
]
